Up version for new pypi (#45)
It seems we missed cutting the v0.4 release on Github.  Thus
when I made the v0.4 not realizing v0.4 was pushed to pypi I created
a naming issue.  Pypi releases must have unique versions.

My solution is to uplevel everything to v0.5 such that the new pypi
version will be 0.1 + 0.4 (the current pypi version).